+++ CHANGELOG.md
@@ -4,9 +4,14 @@
     * Action to paste as a composition
     * Action to import animations
     * Added checks to avoid cyclical dependencies between compositions
+    * Context menu action to "Decompose" PreComp layers
 * UI:
     * Right clicking on the layer view to get a context menu no longer changes the selection
     * Object context menus now have actions to toggle visibility and lock
+* Scripting:
+    * Python has now access to the users of an asset
+* Bug Fixes:
+    * Duplicating shapes updates uuids
 
 ## 0.3.2